data_pipeline_local_original/pyproject.toml
[project]
name = "rt-data-pipeline-local"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = []

[dependency-groups]
producer = [
  "kafka-python==2.3.0",
  "faker==40.1.0",
  "click==8.3.1",
  "pydantic==2.12.5",
  "python-dotenv==1.2.1",
  "pyyaml==6.0.3",
]
flink = [
  "apache-flink==1.18.1 ; python_version < '3.13'",
  "pyyaml==6.0.3",
  "python-dotenv==1.2.1",
  "pydantic==2.12.5",
]
test = [
  "pytest==9.0.2",
  "pytest-timeout==2.4.0",
  "kafka-python==2.3.0",
]
